fix: multiple issues due duplicate useCertificadosFichas.ts file
