+++ app/admin/productos/page.tsx
@@ -1,0 +1,10 @@
+'use client'
+
+export default function ProductosPage() {
+  return (
+    <div className="container mx-auto p-4">
+      <h1 className="text-2xl font-bold mb-6">Productos</h1>
+      <p>Página de productos en desarrollo...</p>
+    </div>
+  )
+}
